fix: initialize m_cacheUsage
- use signed int to prevent potential negative overflow

+++ core/src/data/dataSource.h
@@ -84,8 +84,8 @@ protected:
 
     CacheMap m_cacheMap;
     CacheList m_cacheList;
-    size_t m_cacheUsage;
-    size_t m_cacheMaxUsage;
+    int m_cacheUsage;
+    int m_cacheMaxUsage;
 
 };
